requirements.txt
# core
fastapi==0.115.6
uvicorn==0.32.1
jinja2==3.1.4

# langchain
langchain==0.3.25
langchain-community==0.3.25
langchain-core>=0.3.65,<1.0.0
langchain-text-splitters>=0.3.8,<1.0.0

# vector / embeddings
chromadb==0.4.24
sentence-transformers==2.7.0
numpy<2.0.0
# tools
requests==2.32.3
numexpr==2.10.1
python-dotenv==1.0.1